+++ R/save_piece_images.R
@@ -1,5 +1,3 @@
-#' @import grid
-
 COMPONENT_AND_SIDES <- c("tile_back", "tile_face", "coin_back", "coin_face",
            "die_face", "suitdie_face",
            "pawn_face", "pawn_back", "belt_face",  "saucer_face", "saucer_back",
@@ -21,17 +19,25 @@ piece_device <- function(filename, piece_side=NULL, cfg=list(), angle=0, suit=1,
 }
 
 pp_device <- function(filename, width, height, res=72, bg="transparent") {
+    args <- list(filename = filename, width = width, height = height,
+                 units = "in", res = res, bg = bg)
+    dev_fn <- pp_device_fn(filename)
+    args <- args[names(args) %in% names(formals(dev_fn))]
+    do.call(dev_fn, args)
+}
+
+pp_device_fn <- function(filename) {
     format <- tools::file_ext(filename)
     switch(format,
-           bmp = grDevices::bmp(filename, width, height, "in", res=res, bg=bg),
-           jpeg = grDevices::jpeg(filename, width, height, "in", res=res, bg=bg),
-           jpg = grDevices::jpeg(filename, width, height, "in", res=res, bg=bg),
-           pdf = grDevices::cairo_pdf(filename, width, height, bg=bg),
-           png = grDevices::png(filename, width, height, "in", res=res, bg=bg),
-           ps = grDevices::cairo_ps(filename, width, height, bg=bg),
-           svg = grDevices::svg(filename, width, height, bg=bg),
-           svgz = grDevices::svg(filename, width, height, bg=bg),
-           tiff = grDevices::tiff(filename, width, height, "in", res=res, bg=bg))
+           bmp = grDevices::bmp,
+           jpeg = grDevices::jpeg,
+           jpg = grDevices::jpeg,
+           pdf = grDevices::cairo_pdf,
+           png = grDevices::png,
+           ps = grDevices::cairo_ps,
+           svg = grDevices::svg,
+           svgz = grDevices::svg,
+           tiff = grDevices::tiff)
 }
 
 piece_filename <- function(directory, piece_side, format, angle,
@@ -66,7 +72,8 @@ save_piece_images <- function(cfg = getOption("piecepackr.cfg", pp_cfg()),
     current_dev <- grDevices::dev.cur()
     if (current_dev > 1) on.exit(grDevices::dev.set(current_dev))
 
-    if (!dir.exists(directory)) stop(paste("directory", directory, "does not exist"))
+    stopifnot(dir.exists(directory))
+
     for (f in format) {
         for (a in angle) make_images_helper(directory, cfg, f, a)
     }