+++ R/save_piece_obj.R
@@ -84,7 +84,7 @@ write_mtl <- function(mtl_filename, png_filename) {
 # 1,2,3,4 -> 1,4,3,2
 rev_shift <- function(x) c(x[1], rev(x[-1]))
 
-# 2-sided token angle
+# (2-sided token) rotation matrix to move from "face" to `side`
 side_R <- function(side) {
     switch(side,
            back = R_y(180),
@@ -92,6 +92,16 @@ side_R <- function(side) {
            left = R_y(-90) %*% R_z(-90),
            right = R_y(90) %*% R_z(90),
            top = R_x(90) %*% R_z(180),
+           diag(3))
+}
+# (2-sided token) rotation matrix to return to "face" from `side`
+side_R_rev <- function(side) {
+    switch(side,
+           back = R_y(-180),
+           base = R_x(90),
+           left = R_z(90) %*% R_y(90),
+           right = R_z(-90) %*% R_y(-90),
+           top = R_z(-180) %*% R_x(-90),
            diag(3))
 }
 
@@ -249,7 +259,7 @@ save_peg_doll_obj <- function(piece_side = "pawn_top", suit = 1, rank = 1,
     width <- whd$width
     height <- whd$height
     depth <- whd$depth
-    if (!nigh(width, depth)) warning("Base of peg doll is not circular")
+    if (!nigh(width, depth)) warn("Base of peg doll is not circular")
     # base
     circle <- Point2D$new(pp_shape("convex72", back=TRUE)$npc_coords)$translate(x = -0.5, y = -0.5)
     xyz_base <- as.data.frame(Point3D$new(x=circle$x, y=-0.5, z=circle$y))